+++ core/class/BaseOptimize.class.php
@@ -46,4 +46,24 @@ class BaseOptimize
     {
         return self::$bestScore;
     }
+
+    /**
+     * Test si Jeedom peut exécuter des actions demandant des privilèges.
+     *
+     * @return bool true si c'est possible
+     */
+    public function canSudo()
+    {
+        return jeedom::isCapable('sudo');
+    }
+
+    /**
+     * Obtenir le répertoire racine de Jeedom
+     *
+     * @return bool|string Chemin de la racine de Jeedom
+     */
+    public function getJeedomRootDirectory()
+    {
+        return realpath(dirname(__FILE__) . '/../../../../');
+    }
 }